Add @zambezi scope for independent release
Until (or unless, prehaps) babel/babel#4766 gets merged we need
to release the UMD transform plugin independently to reap the
benefits of these changes. To do this, we add the @zambezi scope
and just release under that.

If and when babel/babel#4766 is indeed merged, we should remove
this independent release.